requirements.txt
# Веб-фреймворк и сопутствующие библиотеки
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic==2.5.1
starlette==0.36.3

# HTTP клиенты и асинхронные библиотеки
aiohttp==3.11.16
requests==2.31.0

# Веб-скрейпинг
beautifulsoup4==4.12.2
lxml==4.9.3
trafilatura==1.6.3

# Базы данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
elasticsearch==8.11.1

# Безопасность и аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.4

# Обработка документов
python-docx==1.0.1
Pillow==10.0.1

# Логирование и утилиты
python-dotenv==1.0.0
tenacity==8.2.3
PyYAML==6.0.1

# Для работы с ElasticSearch/векторными БД
numpy==1.26.0
sentence-transformers==2.2.2

# Форматирование и оптимизация кода
black==23.9.1
isort==5.12.0

# Для тестирования
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.1

# Для мониторинга и производительности
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0

# Интеграция с OpenAI/OpenRouter API
openai==1.2.4

# Дополнительные библиотеки для асинхронных сессий
asyncpg==0.29.0
aiofiles==23.2.1
websockets==12.0

itsdangerous==2.2.0
email-validator==2.1.0.post1
python-docx 
PyMuPDF 
pytesseract 
pdf2image 
python-magic-bin 
mammoth
fitz
aiosmtplib==2.0.2
aiosqlite==0.19.0
jinja2==3.1.2